Move components into new folder
