+++ src/main.js
@@ -1,9 +1,9 @@
-import App from './App.svelte';
+import App from './Components/App.svelte';
 
 const app = new App({
 	target: document.body,
 	props: {
-		name: 'world'
+		name: 'synonyms'
 	}
 });
